Changes to allow loading of VSI files:
Can now load a smaller version of the VSI to save the ROI. Loading handles
scaling factors.
Also fixed bug caused by saved rectangular ROIs being loaded as 4-sided
polygons
Add option to open and ROI in a new image (rather than the current one) to
allow finding the ROI coords for using "crop on load" option to
bio-formats. Currently this step is still in progress. It creates a huge
black image which seems computationallly as bad as loading the whole image.
I could really do with a way of reading the ROI without loading it into
an image.